makefile fix+subsection fix

+++ makefile
@@ -36,8 +36,9 @@ book:
 .PHONY: clean
 clean:
 	rm -f *.aux *.bbl *.blg *.log *.fdb_latexmk *.fls *.out *.toc
-	rm -i *.pdf
-	rm -i book.tex
+	if [ -f book.tex ]; then rm -i book.tex; fi
+	if [ -f book.pdf ]; then rm -i book.pdf; fi
+	for f in *.pdf; do if [ -f "$$f" ]; then rm -i *.pdf; fi; done
 
 .PHONY: all
 all: pdfs book
